src/main/java/com/gildedrose/strategies/ConjuredItemStrategy.java
package com.gildedrose.strategies;

import com.gildedrose.Item;
import com.gildedrose.interfaces.ItemStrategy;

public class ConjuredItemStrategy implements ItemStrategy {
    @Override
    public Item updateItem(Item item) {
        int newSellIn = this.isSellInPositive(item.getSellIn()) ? item.getSellIn() - 1 : 0;
        int newQuality = item.getQuality() - 2 > 0 ? item.getQuality() - 2 : 0;

        item.setSellIn(newSellIn);
        item.setQuality(newQuality);

        return item;
    }

    private boolean isSellInPositive(int sellIn) {
        return sellIn > 0;
    }
}
